+++ src/services/Providers.php
@@ -57,6 +57,9 @@ class Providers extends Component
     public function getRegisteredProviders(): array
     {
         $providerTypes = [
+            registeredproviders\Aramex::class,
+            registeredproviders\AramexAustralia::class,
+            registeredproviders\AramexNewZealand::class,
             registeredproviders\AustraliaPost::class,
             registeredproviders\Bring::class,
             registeredproviders\CanadaPost::class,